Move the TrackView include files out of src/gui/include
