+++ src/core/tracks/AutomationTrack.cpp
@@ -26,7 +26,6 @@
 
 #include "AutomationTrack.h"
 
-#include "AutomationTrackView.h"
 #include "AutomationClip.h"
 
 namespace lmms
