+++ src/core/tracks/PatternTrack.cpp
@@ -29,7 +29,6 @@
 #include "Engine.h"
 #include "PatternClip.h"
 #include "PatternStore.h"
-#include "PatternTrackView.h"
 #include "PlayHandle.h"
