+++ src/gui/tracks/AutomationTrackView.h
