+++ src/gui/tracks/InstrumentTrackView.h
