+++ src/gui/tracks/PatternTrackView.h
