+++ src/gui/tracks/SampleTrackView.h
